Moved models and views in Odoo style
