+++ xpr_xis_connector/i18n/fr.po
