+++ xpr_xis_connector/ir.ui.menu.csv
