+++ xpr_xis_connector/security/ir.model.access.csv
