+++ xpr_xis_connector/views/product.xml
