+++ xpr_xis_connector/views/res_partner.xml
