+++ xpr_xis_connector/views/res_user.xml
