+++ xpr_xis_connector/views/sale.xml
